requirements.txt
requests
bs4
spacy
python-telegram-bot[job-queue] --pre
audiofile
librosa
pdfrw
flask
beautifulsoup4 
numpy
scipy
Pillow
matplotlib
pandas
transformers
openai
python-dateutil 
ipykernel
jupyter
scihub
pyzotero
google-cloud-texttospeech
wave
torch
langchain
tiktoken
python-dotenv
crossrefapi
semanticscholar
langchain_experimental
plotly
scikit-learn
PyPDF2
faiss-cpu
tiktoken
pdf2image
pinecone-client
unstructured
pdfminer.six
sentencepiece
rlvoice
gtts
transformers
datasets
torch
soundfile
playsound
selenium
webdriver_manager
criptography